refactor: radioInput -> toggleGroupInput
