+++ packages/react-router-v7-example/app/components/inspector-sidebar/inspectPanel/panels/helpers/toggleGroupInput.tsx
